+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+### Added
+- check-es-indices-field-number.rb: check if the number of fields in index(es) is approaching limit (default to 1000 in ES)
+### Changed
+- Address a couple of rubocop style violations.
 
 ## [1.11.1] - 2018-02-27
 ### Changed